// src/lib/gen/glimway/v1/profile_pb.d.ts
export declare type HabiticaUser = Message<"glimway.v1.HabiticaUser"> & {
  /**
   * @generated from field: string id = 1 [json_name = "_id"];
   */
  id: string;

  /**
   * @generated from field: glimway.v1.HabiticaUserStats stats = 2;
   */
  stats?: HabiticaUserStats | undefined;

  /**
   * @generated from field: glimway.v1.HabiticaUserName profile = 3;
   */
  profile?: HabiticaUserName | undefined;

  /**
   * @generated from field: glimway.v1.HabiticaUserFlags flags = 4;
   */
  flags?: HabiticaUserFlags | undefined;

  /**
   * @generated from field: glimway.v1.HabiticaUserItems items = 5;
   */
  items?: HabiticaUserItems | undefined;

  /**
   * @generated from field: glimway.v1.HabiticaUserPreferences preferences = 6;
   */
  preferences?: HabiticaUserPreferences | undefined;

  /**
   * @generated from field: glimway.v1.HabiticaUserParty party = 7;
   */
  party?: HabiticaUserParty | undefined;
};

export declare type HabiticaUserValid = HabiticaUser;

/**
 * Describes the message glimway.v1.HabiticaUser.
 * Use `create(HabiticaUserSchema)` to create a new message.
 */
export declare const HabiticaUserSchema: GenMessage<HabiticaUser, {validType: HabiticaUserValid}>;

/**
 * @generated from message glimway.v1.HabiticaUserStats
 */
export declare type HabiticaUserStats = Message<"glimway.v1.HabiticaUserStats"> & {
  /**
   * @generated from field: optional double hp = 1;
   */
  hp?: number | undefined;

  /**
   * @generated from field: optional double mp = 2;
   */
  mp?: number | undefined;

  /**
   * @generated from field: optional double lvl = 3;
   */
  lvl?: number | undefined;

  /**
   * @generated from field: optional double exp = 4;
   */
  exp?: number | undefined;

  /**
   * @generated from field: optional string class = 5;
   */
  class?: string | undefined;

  /**
   * @generated from field: optional double str = 6;
   */
  str?: number | undefined;

  /**
   * @generated from field: optional double int = 7;
   */
  int?: number | undefined;

  /**
   * @generated from field: optional double con = 8;
   */
  con?: number | undefined;

  /**
   * @generated from field: optional double per = 9;
   */
  per?: number | undefined;

  /**
   * @generated from field: glimway.v1.HabiticaBuffs buffs = 10;
   */
  buffs?: HabiticaBuffs | undefined;

  /**
   * @generated from field: optional double max_health = 11;
   */
  maxHealth?: number | undefined;

  /**
   * @generated from field: optional double max_mp = 12 [json_name = "maxMP"];
   */
  maxMp?: number | undefined;
};
// ...
export declare type HabiticaUserName = Message<"glimway.v1.HabiticaUserName"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;
};

export declare type HabiticaUserNameValid = HabiticaUserName;

/**
 * Describes the message glimway.v1.HabiticaUserName.
 * Use `create(HabiticaUserNameSchema)` to create a new message.
 */
export declare const HabiticaUserNameSchema: GenMessage<HabiticaUserName, {validType: HabiticaUserNameValid}>;

/**
 * @generated from message glimway.v1.HabiticaUserFlags
 */
export declare type HabiticaUserFlags = Message<"glimway.v1.HabiticaUserFlags"> & {
  /**
   * @generated from field: optional bool class_selected = 1;
   */
  classSelected?: boolean | undefined;
};

export declare type HabiticaUserFlagsValid = HabiticaUserFlags;

/**
 * Describes the message glimway.v1.HabiticaUserFlags.
 * Use `create(HabiticaUserFlagsSchema)` to create a new message.
 */
export declare const HabiticaUserFlagsSchema: GenMessage<HabiticaUserFlags, {validType: HabiticaUserFlagsValid}>;

/**
 * @generated from message glimway.v1.HabiticaUserParty
 */
export declare type HabiticaUserParty = Message<"glimway.v1.HabiticaUserParty"> & {
  /**
   * @generated from field: google.protobuf.StringValue id = 1 [json_name = "_id"];
   */
  id?: string | undefined;
};
// ...
export declare type HabiticaUserItems = Message<"glimway.v1.HabiticaUserItems"> & {
  /**
   * @generated from field: glimway.v1.HabiticaUserGear gear = 1;
   */
  gear?: HabiticaUserGear | undefined;

  /**
   * @generated from field: map<string, google.protobuf.Value> pets = 2;
   */
  pets: { [key: string]: Value };

  /**
   * @generated from field: map<string, google.protobuf.Value> mounts = 3;
   */
  mounts: { [key: string]: Value };

  /**
   * @generated from field: google.protobuf.StringValue current_pet = 4;
   */
  currentPet?: string | undefined;

  /**
   * @generated from field: google.protobuf.StringValue current_mount = 5;
   */
  currentMount?: string | undefined;
};
// ...
export declare type HabiticaUserPreferences = Message<"glimway.v1.HabiticaUserPreferences"> & {
  /**
   * @generated from field: string size = 1;
   */
  size: string;

  /**
   * @generated from field: string shirt = 2;
   */
  shirt: string;

  /**
   * @generated from field: string skin = 3;
   */
  skin: string;

  /**
   * @generated from field: string background = 4;
   */
  background: string;

  /**
   * @generated from field: glimway.v1.HabiticaUserHair hair = 5;
   */
  hair?: HabiticaUserHair | undefined;

  /**
   * Habitica historically returns true or 1 for costume.
   *
   * @generated from field: google.protobuf.Value costume = 6;
   */
  costume?: Value | undefined;
};